lab02a.6 WorkerPool refactoring, Map interface used
