+++ mark/src/main/java/lab02a/workerspool/IWorkRunnable.java
@@ -1,4 +1,4 @@
-package lab02a.common;
+package lab02a.workerspool;
 
 /**
  * Created by Mark Lobanov on 28.09.2016.
@@ -8,4 +8,5 @@ public interface IWorkRunnable {
 
     public void runWork();
 
+    public void setKeyInfo(WorkerKeyInfo keyInfo);
 }
